add un/re do tests
