+++ package.json
@@ -51,7 +51,6 @@
         "clean:tsbuild": "rimraf tsconfig.app.tsbuildinfo tsconfig.node.tsbuildinfo tsconfig.build.tsbuildinfo",
         "clean": "bun clean:lintcache && bun clean:lib && bun clean:coverage && bun clean:out && bun clean:tsbuild",
         "schema": "bun scripts/schema.ts && bun format",
-        "test:components": "vitest run",
         "test:snapshots": "vitest run --browser.enabled --browser.headless tests/browser/** --update",
         "test:ui": "vitest --ui",
         "test:coverage": "vitest run --coverage",
@@ -60,6 +59,8 @@
         "pretest:browser": "bun monaco && rimraf tests/browser/videos",
         "posttest:browser": "bun scripts/video.ts",
         "posttest:snapshots": "bun scripts/video.ts",
+        "test:components": "vitest run tests/components/**",
+        "test:models": "vitest run tests/models/**",
         "test:all": "bun run test:browser && bun run test:coverage",
         "changelog": "bun scripts/changelog.ts",
         "all": "bun clean && bun forlint && bun schema && bun format && bun run test:all && bun run build && bun archive && bun docs"
@@ -119,7 +120,7 @@
         "playwright": "^1.48.2",
         "pre-commit": "^1.2.2",
         "prettier": "^3.3.3",
-        "react-hotkeys-hook": "^4.5.1",
+        "react-hotkeys-hook": "^4.6.1",
         "react-select-event": "^5.5.1",
         "rimraf": "^6.0.1",
         "stylelint": "^16.10.0",
